fix(elf-dashboard): fix issues with adding presents by elves, restrict user to delete presents, that were appointed by another user

+++ src/components/ChildPresentsTable/ChildPresentsTable.tsx
@@ -1,19 +1,15 @@
-import { MouseEventHandler } from "react";
+import { Fragment, MouseEventHandler } from "react";
 import { useToggle } from "../../hooks";
 import {
+  deleteChildAlternativePresent,
   deleteChildPresent,
-  getUserError,
-  getUserIsLoading,
-  getUserServerMessage,
+  getChildInfo,
+  getUser,
   useAppDispatch,
   useAppSelector,
 } from "../../store";
-import { IPresent } from "../../types";
+import { IPresent, USER_ROLES } from "../../types";
 import { AddChildPresentForm } from "../AddChildPresentForm";
-import {
-  NotificationModal,
-  NotificationModalStatus,
-} from "../NotificationModal";
 import {
   AddPresentButton,
   PresentImage,
@@ -29,12 +25,10 @@ interface IProps {
 }
 
 export const ChildPresentsTable = ({ presents }: IProps) => {
-  const [isAddPresentFormOpen, toggleAddPresentForm] = useToggle();
+  const user = useAppSelector(getUser);
+  const childInfo = useAppSelector(getChildInfo);
 
-  const [isModalOpen, toggleModal] = useToggle();
-  const userDeletePresentIsLoading = useAppSelector(getUserIsLoading);
-  const userDeletePresentError = useAppSelector(getUserError);
-  const userDeletePresentMessage = useAppSelector(getUserServerMessage);
+  const [isAddPresentFormOpen, toggleAddPresentForm] = useToggle();
 
   const dispatch = useAppDispatch();
 
@@ -44,8 +38,22 @@ export const ChildPresentsTable = ({ presents }: IProps) => {
     const { currentTarget } = event;
     if (currentTarget.dataset && currentTarget.dataset.present_id) {
       const { present_id } = currentTarget.dataset;
-      dispatch(deleteChildPresent(Number(present_id)));
-      toggleModal();
+      if (user && user.role === USER_ROLES.Child) {
+        dispatch(deleteChildPresent(Number(present_id)));
+      }
+
+      if (
+        user &&
+        (user.role === USER_ROLES.Elf || user.role === USER_ROLES.DeadMoroz) &&
+        childInfo
+      ) {
+        dispatch(
+          deleteChildAlternativePresent({
+            presentId: Number(present_id),
+            childProfileId: childInfo?.child.profileId,
+          })
+        );
+      }
     }
   };
 
@@ -65,25 +73,27 @@ export const ChildPresentsTable = ({ presents }: IProps) => {
       </thead>
       <tbody>
         {presents ? (
-          presents.map(({ id, name, image }, idx) => {
+          presents.map(({ id, name, image, createdUserId }, idx) => {
             return (
-              <>
-                <tr key={name}>
+              <Fragment key={name}>
+                <tr>
                   <TableData scope="row">{idx + 1}</TableData>
                   <TableData>{name}</TableData>
                   <TableData>
                     {image.url ? <PresentImage src={image.url} /> : "No image"}
                   </TableData>
                   <TableData>
-                    <DeletePresentButton
-                      data-present_id={id}
-                      onClick={deleteChildPresentOnClick}
-                    >
-                      <DeletePresentIcon />
-                    </DeletePresentButton>
+                    {user?.id === createdUserId ? (
+                      <DeletePresentButton
+                        data-present_id={id}
+                        onClick={deleteChildPresentOnClick}
+                      >
+                        <DeletePresentIcon />
+                      </DeletePresentButton>
+                    ) : null}
                   </TableData>
                 </tr>
-              </>
+              </Fragment>
             );
           })
         ) : (
@@ -103,22 +113,6 @@ export const ChildPresentsTable = ({ presents }: IProps) => {
           </TableData>
         </tr>
       </tfoot>
-      {!userDeletePresentIsLoading && (
-        <NotificationModal
-          isOpen={isModalOpen}
-          status={
-            userDeletePresentError
-              ? NotificationModalStatus.Error
-              : NotificationModalStatus.Success
-          }
-          message={
-            userDeletePresentError
-              ? userDeletePresentError
-              : userDeletePresentMessage || "Oooooooooooooops"
-          }
-          handler={toggleModal}
-        />
-      )}
     </StyledChildPresentsTable>
   );
 };
